Move Ruby access paths functions
