+++ extensions/ql-vscode/src/model-editor/languages/ruby/index.ts
@@ -4,48 +4,13 @@ import { Mode } from "../../shared/mode";
 import { parseGenerateModelResults } from "./generate";
 import type { MethodArgument } from "../../method";
 import { getArgumentsList } from "../../method";
-
-function parseRubyMethodFromPath(path: string): string {
-  const match = path.match(/Method\[([^\]]+)].*/);
-  if (match) {
-    return match[1];
-  } else {
-    return "";
-  }
-}
-
-function parseRubyAccessPath(path: string): {
-  methodName: string;
-  path: string;
-} {
-  const match = path.match(/Method\[([^\]]+)]\.(.*)/);
-  if (match) {
-    return { methodName: match[1], path: match[2] };
-  } else {
-    return { methodName: "", path: "" };
-  }
-}
-
-function rubyMethodSignature(typeName: string, methodName: string) {
-  return `${typeName}#${methodName}`;
-}
-
-function rubyMethodPath(methodName: string) {
-  if (methodName === "") {
-    return "";
-  }
-
-  return `Method[${methodName}]`;
-}
-
-function rubyPath(methodName: string, path: string) {
-  const methodPath = rubyMethodPath(methodName);
-  if (methodPath === "") {
-    return path;
-  }
-
-  return `${methodPath}.${path}`;
-}
+import {
+  parseRubyAccessPath,
+  parseRubyMethodFromPath,
+  rubyMethodPath,
+  rubyMethodSignature,
+  rubyPath,
+} from "./access-paths";
 
 export const ruby: ModelsAsDataLanguage = {
   availableModes: [Mode.Framework],
